refactor(statements): extract brokerage import-payload helpers to a service
The statements router inlined brokerage-import domain logic: building/recovering
the import payload from a parsed statement, its transactions, and persisted OCR
metadata, plus the not-ready-reason message (lines 595-680). Move the five
helpers to services/brokerage_statement_payload.py; the router imports the three
it calls. The risky upload_statement orchestration (storage/DB/pipeline side
effects) is intentionally left for a separate focused PR.

routers/statements.py: 947 -> 867 lines. Behavior unchanged; routes keep paths/
signatures; the one test importing _brokerage_payload_from_statement is
repointed. ruff clean; 130 statements + brokerage tests green.
