+++ apps/backend/src/routers/statements.py
@@ -24,7 +24,6 @@ from src.models import (
     BankStatementStatus,
     UploadedDocument,
 )
-from src.models.layer2 import AtomicTransaction
 from src.models.statement_summary import StatementSummary
 from src.schemas import (
     AtomicTransactionResponse,
@@ -47,6 +46,11 @@ from src.schemas.review import (
 from src.services import StorageError, StorageService
 from src.services.ai_models import ModelCatalogError, get_model_info, model_matches_modality
 from src.services.brokerage_positions import BrokeragePositionImportService
+from src.services.brokerage_statement_payload import (
+    _brokerage_import_not_ready_reason,
+    _brokerage_payload_from_persisted_extraction,
+    _brokerage_payload_from_statement,
+)
 from src.services.statement_pipeline import submit_parse_pipeline
 from src.services.statement_posting import auto_create_posted_entries_for_statement, resolve_statement_posting_account
 from src.services.statement_validation import (
@@ -592,94 +596,6 @@ async def list_statement_transactions(
     return BankStatementTransactionListResponse(items=items, total=len(items))
 
 
-def _brokerage_payload_from_statement(
-    statement: StatementSummary,
-    transactions: list[AtomicTransaction],
-) -> dict:
-    """Build an import payload from a parsed brokerage statement."""
-    events = []
-    for txn in transactions:
-        direction = txn.direction.value if hasattr(txn.direction, "value") else txn.direction
-        signed_amount = txn.amount if direction == "IN" else -txn.amount
-        events.append(
-            {
-                "date": txn.txn_date.isoformat(),
-                "description": txn.description,
-                "amount": str(signed_amount),
-                "currency": txn.currency or statement.currency,
-                "raw_text": txn.description,
-            }
-        )
-
-    return {
-        "institution": statement.institution,
-        "statement": {
-            "institution": statement.institution,
-            "period_end": statement.period_end.isoformat() if statement.period_end else None,
-            "currency": statement.currency,
-        },
-        "transactions": events,
-        "events": events,
-    }
-
-
-def _extract_brokerage_payload_from_metadata(metadata: dict | None) -> dict | None:
-    """Return the structured extraction payload stored in Layer 1 metadata."""
-    if not isinstance(metadata, dict):
-        return None
-    for key in ("extraction_payload", "parsed_payload", "payload"):
-        payload = metadata.get(key)
-        if isinstance(payload, dict):
-            return payload
-    return metadata if any(key in metadata for key in ("positions", "holdings", "securities")) else None
-
-
-def _enrich_brokerage_payload_from_statement(payload: dict, statement: StatementSummary) -> dict:
-    """Backfill statement metadata into a recovered extraction payload."""
-    enriched = dict(payload)
-    enriched.setdefault("institution", statement.institution)
-    statement_payload = enriched.get("statement") if isinstance(enriched.get("statement"), dict) else {}
-    statement_payload = dict(statement_payload)
-    statement_payload.setdefault("institution", statement.institution)
-    statement_payload.setdefault("period_end", statement.period_end.isoformat() if statement.period_end else None)
-    statement_payload.setdefault("currency", statement.currency)
-    enriched["statement"] = statement_payload
-    return enriched
-
-
-async def _brokerage_payload_from_persisted_extraction(
-    db,
-    *,
-    statement: StatementSummary,
-    uploaded_document: UploadedDocument | None,
-    user_id: UUID,
-) -> dict | None:
-    """Load the persisted OCR extraction payload for statement-scoped imports."""
-    payload = _extract_brokerage_payload_from_metadata(statement.extraction_metadata)
-    if payload is not None:
-        return _enrich_brokerage_payload_from_statement(payload, statement)
-
-    if uploaded_document is None:
-        return None
-    payload = _extract_brokerage_payload_from_metadata(uploaded_document.extraction_metadata)
-    if payload is None:
-        return None
-    return _enrich_brokerage_payload_from_statement(payload, statement)
-
-
-def _brokerage_import_not_ready_reason(statement: StatementSummary, transaction_count: int) -> str:
-    """Explain why a brokerage statement cannot be imported yet."""
-    status_value = statement.status.value if hasattr(statement.status, "value") else str(statement.status)
-    validation_error = statement.validation_error
-
-    if status_value == BankStatementStatus.REJECTED.value:
-        return f"Provider parsing failed before brokerage import: {validation_error or 'statement rejected'}"
-    if status_value in {BankStatementStatus.UPLOADED.value, BankStatementStatus.PARSING.value}:
-        return "Provider parsing has not completed; statement must be parsed before brokerage import"
-
-    return f"Statement must be parsed before importing brokerage positions; current status={status_value}"
-
-
 # Synchronous 200: the import runs inline and returns the full BrokerageImportResponse
 # (counts + reconciliation results), so the operation is complete on response — not a
 # 202 background job (cf. #1099 AC12.29.1).
@@ -703,9 +619,7 @@ async def import_brokerage_statement_positions(
     uploaded_document = await _resolve_uploaded_document(db, statement, user_id)
     source_filename = uploaded_document.original_filename if uploaded_document else statement.file_hash
 
-    payload = await _brokerage_payload_from_persisted_extraction(
-        db, statement=statement, uploaded_document=uploaded_document, user_id=user_id
-    )
+    payload = _brokerage_payload_from_persisted_extraction(statement=statement, uploaded_document=uploaded_document)
     if payload is None:
         payload = _brokerage_payload_from_statement(statement, transactions)
     request_id = _current_request_id()
